Development

+++ Profitable/Profitable.Services/Positions/PositionsRecordsService.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.EntityFrameworkCore;
 using Profitable.Common.Enums;
+using Profitable.Common.GlobalConstants;
 using Profitable.Common.Models;
 using Profitable.Common.Services;
 using Profitable.Data.Repository.Contract;
@@ -52,46 +53,70 @@ namespace Profitable.Services.Positions
 
 		public async Task<Result> ChangeNamePositionsRecordList(
 			Guid recordGuid,
-			string recordName)
+			string recordName,
+			Guid requesterGuid)
 		{
-			var recordToUpdate = await repository
+            try
+            {
+                var recordToUpdate = await repository
 				.GetAllAsNoTracking()
 				.Where(post => !post.IsDeleted)
 				.FirstOrDefaultAsync(entity => entity.Guid == recordGuid);
 
-			if (recordToUpdate != null)
-			{
-				recordToUpdate.Name = recordName;
+				if (recordToUpdate == null)
+				{
+					return GlobalServicesConstants.EntityDoesNotExist("Positions record");
+				}
+			
+				if(recordToUpdate.UserId != requesterGuid)
+				{
+					return GlobalServicesConstants.RequesterNotOwnerMesssage;
+				}
+			
+                recordToUpdate.Name = recordName;
 
-				repository.Update(recordToUpdate);
+                repository.Update(recordToUpdate);
+
+                await repository.SaveChangesAsync();
+
+                return true;
+            }
+			catch (Exception e)
+			{
+				return e.Message;
+			}
+        }
+
+		public async Task<Result> DeletePositionsRecordList(Guid recordGuid, Guid requesterGuid)
+		{
+            try
+            {
+                var record = await repository
+                .GetAllAsNoTracking()
+                .FirstOrDefaultAsync(entity => entity.Guid == recordGuid);
+
+				if (record == null)
+				{
+					return GlobalServicesConstants.EntityDoesNotExist("Positions record");
+				}
+
+				if (record.UserId != requesterGuid)
+				{
+					return GlobalServicesConstants.RequesterNotOwnerMesssage;
+				}
+
+				repository.Delete(record);
 
 				await repository.SaveChangesAsync();
 
 				return true;
+
 			}
-			else
+			catch (Exception e)
 			{
-				return "Positions Record was not found!";
+				return e.Message;
 			}
 		}
-
-		public async Task<Result> DeletePositionsRecordList(Guid recordGuid)
-		{
-            var record = await repository
-                .GetAllAsNoTracking()
-                .FirstOrDefaultAsync(entity => entity.Guid == recordGuid);
-
-            if (record == null)
-            {
-                return "Record not found";
-            }
-
-            repository.Delete(record);
-
-            await repository.SaveChangesAsync();
-
-            return true;
-        }
 
 		public IEnumerable<string> GetPositionsRecordsOrderTypes()
 		{
